CDMdbHandler

Class for handling database connection and schema information for a CDM database

Active bindings

databaseId

A text id for the database the it connects to (read-only).

databaseName

A text id for the database the it connects to (read-only).

databaseDescription

A text description for the database the it connects to (read-only).

connectionHandler

ConnectionHandler object for managing the database connection (read-only).

vocabularyDatabaseSchema

Name of the vocabulary database schema (read-only).

cdmDatabaseSchema

Name of the CDM database schema (read-only).

connectionStatusLog

Log tibble object for storing connection status information (read-only).

vocabularyInfo

Data frame containing information about the vocabulary database (read-only).

CDMInfo

Data frame containing information about the CDM database (read-only).

getTblVocabularySchema

List of functions that create dbplyr table for the vocabulary tables (read-only).

getTblCDMSchema

List of functions that create dbplyr table for the CDM tables (read-only).

Methods

Public methods

CDMdbHandler$new()

CDMdbHandler$finalize()

CDMdbHandler$loadConnection()

CDMdbHandler$clone()

Method new()

Usage

CDMdbHandler$new(
  databaseId,
  databaseName,
  databaseDescription,
  connectionHandler,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  loadConnectionChecksLevel = "allChecks"
)

Arguments

databaseId

A text id for the database the it connects to

databaseName

A text name for the database the it connects to

databaseDescription

A text description for the database the it connects to

connectionHandler

A ConnectionHandler object

cdmDatabaseSchema

Name of the CDM database schema

vocabularyDatabaseSchema

(Optional) Name of the vocabulary database schema (default is cdmDatabaseSchema)

loadConnectionChecksLevel

(Optional) Level of checks to perform when loading the connection (default is "allChecks") Finalize method

Method finalize()

Closes the connection if active. Reload connection

Usage

CDMdbHandler$finalize()

Method loadConnection()

Updates the connection status by checking the database connection, vocabulary database schema, and CDM database schema.

Usage

CDMdbHandler$loadConnection(loadConnectionChecksLevel)

Method clone()

The objects of this class are cloneable with this method.

Usage

CDMdbHandler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.