Operation Subset Operator

A subset of type cohort - subset a population to only those contained within defined cohort to List

Super class

CohortGenerator::SubsetOperator -> OperationSubsetOperator

Active bindings

operationString

Integer id of cohorts to match to

Methods

Public methods

CohortGenerator_OperationSubsetOperator$toList()

CohortGenerator_OperationSubsetOperator$getAutoGeneratedName()

CohortGenerator_OperationSubsetOperator$clone()

Inherited methods

CohortGenerator::SubsetOperator$classname()

CohortGenerator::SubsetOperator$getQueryBuilder()

CohortGenerator::SubsetOperator$initialize()

CohortGenerator::SubsetOperator$isEqualTo()

CohortGenerator::SubsetOperator$publicFields()

CohortGenerator::SubsetOperator$toJSON()

Method toList()

List representation of object Get auto generated name

Usage

CohortGenerator_OperationSubsetOperator$toList()

Method getAutoGeneratedName()

name generated from subset operation properties

Usage

CohortGenerator_OperationSubsetOperator$getAutoGeneratedName()

Returns

character

Method clone()

The objects of this class are cloneable with this method.

Usage

CohortGenerator_OperationSubsetOperator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.