CohortTableHandler

Class for handling cohort tables in a CDM database. Inherits from CDMdbHandler.

Super class

HadesExtras::CDMdbHandler -> CohortTableHandler

Active bindings

cohortDatabaseSchema

Schema where cohort tables are stored

cohortTableNames

Names of the cohort tables in the database

incrementalFolder

Path to folder for incremental operations

cohortDefinitionSet

Set of cohort definitions

cohortGeneratorResults

Results from cohort generation process

cohortDemograpics

Demographic information for cohorts

cohortsOverlap

Information about overlapping cohorts

Methods

Public methods

CohortTableHandler$new()

CohortTableHandler$finalize()

CohortTableHandler$loadConnection()

CohortTableHandler$insertOrUpdateCohorts()

CohortTableHandler$deleteCohorts()

CohortTableHandler$getCohortCounts()

CohortTableHandler$getCohortsSummary()

CohortTableHandler$getCohortIdAndNames()

CohortTableHandler$updateCohortNames()

CohortTableHandler$getCohortsOverlap()

CohortTableHandler$clone()

Method new()

Initialize the CohortTableHandler object

Usage

CohortTableHandler$new(
  connectionHandler,
  databaseId,
  databaseName,
  databaseDescription,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTableName,
  loadConnectionChecksLevel = "allChecks"
)

Arguments

connectionHandler

The connection handler object.

databaseId

A text id for the database the it connects to

databaseName

A text name for the database the it connects to

databaseDescription

A text description for the database the it connects to

cdmDatabaseSchema

Name of the CDM database schema.

vocabularyDatabaseSchema

Name of the vocabulary database schema. Default is the same as the CDM database schema.

cohortDatabaseSchema

Name of the cohort database schema.

cohortTableName

Name of the cohort table.

loadConnectionChecksLevel

(Optional) Level of checks to perform when loading the connection (default is "allChecks") Finalize method

Method finalize()

Closes the connection if active.

loadConnection

Usage

CohortTableHandler$finalize()

Method loadConnection()

Reloads the connection with the initial setting and updates connection status

insertOrUpdateCohorts

Usage

CohortTableHandler$loadConnection(loadConnectionChecksLevel)

Arguments

loadConnectionChecksLevel

Level of checks to perform during connection

Method insertOrUpdateCohorts()

If there is no cohort with the same cohortId it is added to the cohortDefinitionSet, If there is a cohort with the same cohortId, the cohort is updated in the cohortDefinitionSet CohortDefinitionSet is generated and demographics is updated for only the cohorts that have changed

Usage

CohortTableHandler$insertOrUpdateCohorts(cohortDefinitionSet)

Arguments

cohortDefinitionSet

The cohort definition set to add.

deleteCohorts

Method deleteCohorts()

Deletes cohorts from the cohort table.

Usage

CohortTableHandler$deleteCohorts(cohortIds)

Arguments

cohortIds

The cohort ids to delete.

getCohortCounts

Method getCohortCounts()

Retrieves cohort counts from the cohort table.

Usage

CohortTableHandler$getCohortCounts()

Returns

A tibble containing the cohort counts with names.

getCohortsSummary

Method getCohortsSummary()

Retrieves the summary of cohorts including cohort start and end year histograms and sex counts.

Usage

CohortTableHandler$getCohortsSummary()

Returns

A tibble containing cohort summary.

getCohortIdAndNames

Method getCohortIdAndNames()

Retrieves the cohort names.

Usage

CohortTableHandler$getCohortIdAndNames()

Returns

A vector with the name of the cohorts

updateCohortNames

Method updateCohortNames()

Updates the cohort name and short name.

Usage

CohortTableHandler$updateCohortNames(cohortId, newCohortName, newShortName)

Arguments

cohortId

The cohort id to update.

newCohortName

New name to assign to the cohort

newShortName

New short name to assign to the cohort

cohortName

The new cohort name.

shortName

The new short name.

getCohortsOverlap

Method getCohortsOverlap()

Retrieves the number of subjects that are in more than one cohort.

Usage

CohortTableHandler$getCohortsOverlap()

Returns

A tibble containing one logical column for each cohort with name a cohort id, and an additional column numberOfSubjects with the number of subjects in the cohorts combination.

Method clone()

The objects of this class are cloneable with this method.

Usage

CohortTableHandler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.